Rename FooAbstract to AbstractFoo [#17]
